+++ desktop/modules/live2d-5/preload.ts
@@ -1,0 +1,17 @@
+/**
+ * Live2D Cubism 5 宠物窗口 Preload 脚本
+ * 桥接主进程与宠物窗口 renderer 的 IPC 通信
+ */
+import { contextBridge, ipcRenderer } from 'electron'
+
+contextBridge.exposeInMainWorld('electronAPI', {
+  invoke: (channel: string, ...args: unknown[]) => ipcRenderer.invoke(channel, ...args),
+  on: (channel: string, callback: (...args: unknown[]) => void) => {
+    const subscription = (_event: Electron.IpcRendererEvent, ...args: unknown[]) => callback(...args)
+    ipcRenderer.on(channel, subscription)
+    return () => { ipcRenderer.removeListener(channel, subscription) }
+  },
+  removeListener: (channel: string, callback: (...args: unknown[]) => void) => {
+    ipcRenderer.removeListener(channel, callback)
+  }
+})